requirements.txt
jira>=3.5.1
openai>=1.0.0
python-dotenv>=0.19.0
rich>=10.0.0
typer>=0.4.0
requests>=2.31.0
nltk>=3.8.1
pandas>=2.0.0
scikit-learn>=1.3.0
python-telegram-bot>=20.0
apscheduler>=3.10.0 